Add unittest for sst

+++ tests/test_file_writer.cpp
@@ -17,9 +17,9 @@ namespace smallkv {
         std::string data1 = "123gfds4h6.1 s0 3ds 4g00 x_";
         std::string data2 = "  7112gdf455 4 56 u 455u 123fgh12fgn  ";
         std::string data3 = ".*-123 34fe r65 4r1 0m1 j0.10,/,.";
-        fwriter->append(data1.data(), data1.size());
-        fwriter->append(data2.data(), data2.size());
-        fwriter->append(data3.data(), data3.size());
+        fwriter->append(data1.data(), static_cast<int32_t>(data1.size()));
+        fwriter->append(data2.data(), static_cast<int32_t>(data2.size()));
+        fwriter->append(data3.data(), static_cast<int32_t>(data3.size()));
         fwriter->sync();
         fwriter->close();
         auto _ = std::system(("rm -rf " + path).c_str());
